pyama-core/pyproject.toml
[project]
name = "pyama-core"
version = "0.1.0"
description = "Core processing library for PyAMA"
authors = [{ name = "PyAMA Team" }]
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "scipy",
    "xarray",
    "bioio",
    "bioio-nd2",
    "bioio-czi",
    "bioio-tifffile",
    "pandas",
    "scikit-image",
    "pyyaml",
    "pylibczirw>=5.0.0",
    "typer",
    "tqdm",
    "cellpose",
    "pytest",
    "matplotlib",
    "spotiflow",
    "btrack",
    # FastAPI server dependencies
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "httpx",  # For testing FastAPI
    # MCP server dependencies
    "mcp>=1.0.0",
]

[tool.uv]
package = true

[[tool.uv.index]]
name = "scm"
url = "https://pypi.scm.io/simple/"

[tool.uv.sources]
pylibczirw = { index = "scm" }

[project.scripts]
pyama-core = "pyama_core.cli.main:app"
